ia/requirements.txt
Flask
flask-cors
kafka-python
pandas
numpy
joblib
scikit-learn
xgboost
requests
